\name{excel_orig}
\alias{excel_orig}
\docType{data}
\title{
Texts to label
}
\description{
This is an illustrative data frame with three posts and all the columns
related to sentiments, emotions and Basic User Aspects (BUAs) to be filled in. 
Specifically, the first column is the text identifier. The second column contains the texts. The third column contains the language in which each text was written. The rest of the columns are empty columns.
}
\usage{excel_orig}
\format{Data frame with 3 rows and 16 columns.}
\source{\url{https://www.symanto.net/}}
\keyword{datasets}
