% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bua.R
\name{get_bua}
\alias{get_bua}
\title{Understand Basic User Aspects (BUA)}
\usage{
get_bua(text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{text}{String with the text.}

\item{language}{String with the language. For the time being, the supported 
language is "en" (English).}

\item{all}{Boolean. If TRUE, all the BUA probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with columns:
\enumerate{
  \item text Text analyzed.
  \item bua BUAs that the text conveys.
  \item probab Probability of the emotion returned.
}
}
\description{
Understanding of the basic user aspects of the author. This is an aggregation call 
for all the other BUA function calls. The BUAs include personality, self revealing,
information seeking, action seeking and fact oriented.
}
\details{
\itemize{
  \item The personality traits are \emph{emotional} or \emph{rational}.
  \item The self revealing checks whether the author shares personal
        feelings (\emph{yes}) or not (\emph{no}).
  \item The information seeking checks whether the author shows an active 
        interest in a subject (\emph{yes}) or not (\emph{no}).
  \item The action seeking checks whether the author seeks to create an 
        impact (\emph{yes}) or not (\emph{no}).
  \item The fact oriented checks whether the author expresses 
        objective facts (\emph{yes}) or not (\emph{no}).
}
}
\examples{
\dontrun{
token <- "Request your API Key: https://developer.symanto.net/"
# English:
text <- "I am very happy"
language <- "en"
get_bua(text, language, all = TRUE, "UTF-8", token)
get_bua(text, language, all = FALSE, "UTF-8", token)
}

}
\seealso{
\code{\link{get_bua_action}}, \code{\link{get_bua_fact}},
\code{\link{get_bua_self}}, \code{\link{get_bua_personality}},
\code{\link{get_bua_info}}
}
\author{
Symanto
}
