% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bua_action.R
\name{get_bua_action}
\alias{get_bua_action}
\title{Action seeking}
\usage{
get_bua_action(text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{text}{String with the text.}

\item{language}{String with the language. For the time being, the supported 
languages are "en" (English) and "de" (German).}

\item{all}{Boolean. If TRUE, all the BUA probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with columns:
\enumerate{
  \item text Text analyzed.
  \item action Checks whether the author seeks to create an 
  impact (\emph{yes}) or not (\emph{no}).
  \item probab Probability of seeking an action.
}
}
\description{
Analysis of the author's intention to create an impact or cause effect 
or to exert an influence. Through a request, a direct opinion or an 
indirect suggestion, the author expects their message to prompt action.
}
\examples{
\dontrun{
token <- "Request your API Key: https://developer.symanto.net/"
# English:
text <- "We cannot accept the government decisions. We must do something."
language <- "en"
get_bua_action(text, language, TRUE, "UTF-8", token)
get_bua_action(text, language, FALSE, "UTF-8", token)

# German:
text <- "Die Regierungsentscheidungen sind inakzeptabel. Man muss etwas machen."
language <- "de"
get_bua_action(text, language, TRUE, "UTF-8", token)
}

}
\seealso{
\code{\link{get_bua}}
}
\author{
Symanto
}
