% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bua_fact.R
\name{get_bua_fact}
\alias{get_bua_fact}
\title{Fact oriented}
\usage{
get_bua_fact(text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{text}{String with the text.}

\item{language}{String with the language. For the time being, the supported 
languages are "en" (English) and "de" (German).}

\item{all}{Boolean. If TRUE, all the BUA probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with columns:
\enumerate{
  \item text Text analyzed.
  \item fact Checks whether the author expresses 
  objective facts (\emph{yes}) or not (\emph{no}).
  \item probab Probability of factual information.
}
}
\description{
Analysis of the author's intention to express objective and factual information.
}
\examples{
\dontrun{
token <- "Request your API Key: https://developer.symanto.net/"
# English:
text <- "Messi is the world's best soccer player."
language <- "en"
get_bua_fact(text, language, TRUE, "UTF-8", token)
get_bua_fact(text, language, FALSE, "UTF-8", token)

# German:
text <- "Messi ist der beste Fussballspieler der Welt."
language <- "de"
get_bua_fact(text, language, TRUE, "UTF-8", token)
}

}
\seealso{
\code{\link{get_bua}}
}
\author{
Symanto
}
