% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bua_info.R
\name{get_bua_info}
\alias{get_bua_info}
\title{Information seeking}
\usage{
get_bua_info(text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{text}{String with the text.}

\item{language}{String with the language. For the time being, the supported 
languages are "en" (English) and "de" (German).}

\item{all}{Boolean. If TRUE, all the BUA probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with columns:
\enumerate{
  \item text Text analyzed.
  \item info Checks whether the author shows an active 
  interest in a subject (\emph{yes}) or not (\emph{no}).
  \item probab Probability of information seeking.
}
}
\description{
Analysis of the author's intention to show an active interest in a subject. 
The message aims to extend knowledge through filling some knowledge gap, 
receive a reply or reassurance to the communication.
}
\note{
The information requested is not limited to facts.
}
\examples{
\dontrun{
token <- "Request your API Key: https://developer.symanto.net/"
# English:
text <- "I am interested in good literature."
language <- "en"
get_bua_info(text, language, TRUE, "UTF-8", token)
get_bua_info(text, language, FALSE, "UTF-8", token)

# German:
text <- "Ich habe ein Interesse an guter Literatur."
language <- "de"
get_bua_info(text, language, TRUE, "UTF-8", token)
}

}
\seealso{
\code{\link{get_bua}}
}
\author{
Symanto
}
