% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bua_personality.R
\name{get_bua_personality}
\alias{get_bua_personality}
\title{Personality}
\usage{
get_bua_personality(text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{text}{String with the text.}

\item{language}{String with the language. For the time being, the supported 
languages are "en" (English) and "de" (German).}

\item{all}{Boolean. If TRUE, all the BUA probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with columns:
\enumerate{
  \item text Text analyzed.
  \item personality Checks whether the personality traits are 
  \emph{emotional} or \emph{rational}.
  \item probab Probability of personality traits.
}
}
\description{
Personality traits evaluate the content and writing style to determine how the 
author makes decisions. The model identifies whether an author is emotional 
(relationship-oriented, focusing on social values and empathy) or rational 
(objective and pragmatic, focusing on facts and logical deduction).
}
\examples{
\dontrun{
token <- "Request your API Key: https://developer.symanto.net/"
# English:
text <- "People get bored at home."
language <- "en"
get_bua_personality(text, language, TRUE, "UTF-8", token)
get_bua_personality(text, language, FALSE, "UTF-8", token)

# German:
text <- "Menschen langweilen sich zu Hause."
language <- "de"
get_bua_personality(text, language, TRUE, "UTF-8", token)
}

}
\seealso{
\code{\link{get_bua}}
}
\author{
Symanto
}
