% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bua_self.R
\name{get_bua_self}
\alias{get_bua_self}
\title{Self revealing}
\usage{
get_bua_self(text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{text}{String with the text.}

\item{language}{String with the language. For the time being, the supported 
languages are "en" (English) and "de" (German).}

\item{all}{Boolean. If TRUE, all the BUA probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with columns:
\enumerate{
  \item text Text analyzed.
  \item self Checks whether the author shares personal
  feelings (\emph{yes}) or not (\emph{no}).
  \item probab Probability of self revealing.
}
}
\description{
Analysis of the author's intention to share personal preferences, 
experiences and feelings.
}
\examples{
\dontrun{
token <- "Request your API Key: https://developer.symanto.net/"
# English:
text <- "Joker is a great movie."
language <- "en"
get_bua_self(text, language, TRUE, "UTF-8", token)
get_bua_self(text, language, FALSE, "UTF-8", token)

# German:
text <- "Joker ist ein grossartiger Film."
language <- "de"
get_bua_self(text, language, TRUE, "UTF-8", token)
}

}
\seealso{
\code{\link{get_bua}}
}
\author{
Symanto
}
