% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ekman_emotion.R
\name{get_ekman_emotion}
\alias{get_ekman_emotion}
\title{Ekman's emotions}
\usage{
get_ekman_emotion(url_api, text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{url_api}{URL to the API.}

\item{text}{String with the texts. A set of maximum 512 texts is allowed.}

\item{language}{Language of the text. Supported languages are 
"en" (English), "de" (German) and "es" (Spanish).}

\item{all}{Boolean. If TRUE, all the probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with the texts and the probabilities of the 
psychological features obtained.
}
\description{
For a text written in a given language, this function extracts the Ekman's 
emotions that it conveys. Returned labels are \emph{anger}, \emph{disgust}, 
\emph{fear}, \emph{joy}, \emph{no_emotion}, \emph{sadness} and \emph{surprise}.
}
\note{
The type of array for one text is: \cr
data <- "[{\"id\":\"1\",\"text\":\"love\",\"language\":\"en\"}]" \cr
For more than one text is: \cr
data <- "[{\"id\":\"1\",\"text\":\"love\",\"language\":\"en\"}, 
\cr
          {\"id\":\"2\",\"text\":\"hate\",\"language\":\"en\"}]" \cr
(write it in one line). 
In the first case, add a curly bracket after the first box bracket and 
before the last box bracket. In the second case, add the curly brackets
for every set of id, text and language.

An explanation of the Ekman's foundations can be found 
\href{https://www.paulekman.com/universal-emotions/}{here.}
}
\examples{
\dontrun{
token <- "Request your API Key: https://developers.symanto.net/signup"
url_api <- "https://api.symanto.net"

# For only one text:
# English:
text <- "I love the service."
language <- "en"
get_ekman_emotion(url_api, text, language, TRUE, "UTF-8", token)
get_ekman_emotion(url_api, text, language, FALSE, "UTF-8", token)

# German:
text <- "Ich bin sehr zufrieden."
language <- "de"
get_ekman_emotion(url_api, text, language, TRUE, "UTF-8", token)
get_ekman_emotion(url_api, text, language, FALSE, "UTF-8", token)

# Spanish:
text <- "Estoy muy contento."
language <- "es"
get_ekman_emotion(url_api, text, language, TRUE, "UTF-8", token)
get_ekman_emotion(url_api, text, language, FALSE, "UTF-8", token)

# For more than one text:
# English:
df <- data.frame(text = c("love", "hate"), language = c("en", "en"))
text <- df$text
language <- df$language
get_ekman_emotion(url_api, text, language, all = TRUE, "UTF-8", token)
get_ekman_emotion(url_api, text, language, all = FALSE, "UTF-8", token)
}

}
\author{
Symanto
}
