% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_language_detection.R
\name{get_language_detection}
\alias{get_language_detection}
\title{Language detection}
\usage{
get_language_detection(url_api, text, encoding = "UTF-8", token)
}
\arguments{
\item{url_api}{URL to the API.}

\item{text}{String with the texts. A set of maximum 64 texts is allowed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with the text and the language in which the text
was written.
}
\description{
This function identifies the language of the input text. 
Any language can be detected.
}
\note{
The type of array for one text is: \cr
data <- "[{\"id\":\"1\",\"text\":\"love\",\"language\":\"en\"}]" \cr
For more than one text is: \cr
data <- "[{\"id\":\"1\",\"text\":\"love\",\"language\":\"en\"}, 
\cr
          {\"id\":\"2\",\"text\":\"hate\",\"language\":\"en\"}]" \cr
(write it in one line). 
In the first case, add a curly bracket after the first box bracket and 
before the last box bracket. In the second case, add the curly brackets
for every set of id, text and language.
}
\examples{
\dontrun{
token <- "Request your API Key: https://developers.symanto.net/signup"
url_api <- "https://api.symanto.net"

# For only one text:
# English:
text <- "Hello I love the service."
get_language_detection(url_api, text, "UTF-8", token)

# German:
text <- "Ich bin sehr zufrieden."
get_language_detection(url_api, text, "UTF-8", token)

# Spanish:
text <- "Estoy muy contento."
get_language_detection(url_api, text, "UTF-8", token)

# For more than one text:
# English:
text <- c("love", "hate", "felicidad")
get_language_detection(url_api, text, "UTF-8", token)
}

}
\author{
Symanto
}
