% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_personality.R
\name{get_personality}
\alias{get_personality}
\title{Personality traits}
\usage{
get_personality(url_api, text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{url_api}{URL to the API.}

\item{text}{String with the texts. A set of maximum 32 texts is allowed.}

\item{language}{Language of the text. Supported languages are
"ar" (Arabic), "de" (German), "en" (English), "es" (Spanish), "fr" (French), 
"it" (Italian), "nl" (Dutch), "pt" (Portuguese), "ru" (Russian), "tr" (Turkish)
 and "zh" (Chinese).}

\item{all}{Boolean. If TRUE, all the probabilities
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with the texts and the probabilities of the 
psychological features obtained.
}
\description{
Personality traits evaluate the content and writing style to determine how the 
author makes decisions. The model identifies whether an author is \emph{emotional}
(relationship-oriented, focusing on social values and empathy) or \emph{rational} 
(objective and pragmatic, focusing on facts and logical deduction).
}
\note{
The type of array for one text is: \cr
data <- "[{\"id\":\"1\",\"text\":\"love\",\"language\":\"en\"}]" \cr
For more than one text is: \cr
data <- "[{\"id\":\"1\",\"text\":\"love\",\"language\":\"en\"}, 
\cr
          {\"id\":\"2\",\"text\":\"hate\",\"language\":\"en\"}]" \cr
(write it in one line). 
In the first case, add a curly bracket after the first box bracket and 
before the last box bracket. In the second case, add the curly brackets
for every set of id, text and language.
}
\examples{
\dontrun{
token <- "Request your API Key: https://developers.symanto.net/signup"
url_api <- "https://api.symanto.net"

# For only one text:
# English:
text <- "I love the service."
language <- "en"
get_personality(url_api, text, language, TRUE, "UTF-8", token)
get_personality(url_api, text, language, FALSE, "UTF-8", token)

# German:
text <- "Menschen langweilen sich zu Hause."
language <- "de"
get_personality(url_api, text, language, TRUE, "UTF-8", token)
get_personality(url_api, text, language, FALSE, "UTF-8", token)

# Spanish:
text <- "Estoy muy contento."
language <- "es"
get_personality(url_api, text, language, TRUE, "UTF-8", token)
get_personality(url_api, text, language, FALSE, "UTF-8", token)

# For more than one text:
# English:
df <- data.frame(text = c("love", "hate"), language = c("en", "en"))
text <- df$text
language <- df$language
get_personality(url_api, text, language, all = TRUE, "UTF-8", token)
get_personality(url_api, text, language, all = FALSE, "UTF-8", token)
}

}
\author{
Symanto
}
