% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiment.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Get sentiment from a text}
\usage{
get_sentiment(text, language, all = FALSE, encoding = "UTF-8", token)
}
\arguments{
\item{text}{String with the text.}

\item{language}{String with the language. Supported languages are 
"en" (English), "de" (German) and "es" (Spanish).}

\item{all}{Boolean. If TRUE, both the positive and negative probabilities 
are displayed. If FALSE, only the largest probability is displayed.}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}
}
\value{
A data frame with columns:
\enumerate{
  \item text Text analyzed.
  \item sentiment Sentiment that the text conveys.
  \item probab Probability of the sentiment returned.
}
}
\description{
For a text written in a given language, this function extracts the sentiment 
it conveys. Returned labels are "positive" and "negative".
}
\examples{
\dontrun{
token <- "Request your API Key: https://developer.symanto.net/"
# English:
text <- "I am very happy"
language <- "en"
get_sentiment(text, language, TRUE, "UTF-8", token)
get_sentiment(text, language, FALSE, "UTF-8", token)

# German:
text <- "Ich bin sehr zufrieden"
language <- "de"
get_sentiment(text, language, TRUE, "UTF-8", token)

# Spanish:
text <- "Estoy muy contento"
language <- "es"
gs <- get_sentiment(text, language, TRUE, "UTF-8", token)
gs
# Probabilities sum indeed one:
sum(gs[grepl("prob", names(gs))])
}

}
\author{
Symanto
}
