% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_text.R
\name{label_text}
\alias{label_text}
\title{Label texts}
\usage{
label_text(excel_file, user_actions, encoding = "UTF-8", token, verbose = FALSE)
}
\arguments{
\item{excel_file}{Excel file.}

\item{user_actions}{Vector with the set of columns (psychological properties) 
that the user wants to fill in. Possible values are "sentiments", "emotions", 
"action_seeking", "fact_oriented", "self_revealing", "personality" and 
"information_seeking".}

\item{encoding}{Character encoding. Default "UTF-8".}

\item{token}{The access token.}

\item{verbose}{Provides a flow of information.}
}
\value{
The \code{excel_file} filled with the probablities related to the 
psychological properties desired by the user.
}
\description{
Given a certain (Excel) file with some posts and all the empty columns related 
to sentiments, emotions and Basic User Aspects (BUAs), this function
fills the columns that the user wants to.
}
\details{
\itemize{
  \item The sentiments are \emph{positive} and \emph{negative}.
  \item The emotions are \emph{anger}, \emph{joy}, \emph{love}, 
  \emph{sadness} and \emph{unrecognized}.
  \item The personality traits are \emph{emotional} or \emph{rational}.
  \item The self revealing checks whether the author shares personal
        feelings (\emph{yes}) or not (\emph{no}).
  \item The information seeking checks whether the author shows an active 
        interest in a subject (\emph{yes}) or not (\emph{no}).
  \item The action seeking checks whether the author seeks to create an 
        impact (\emph{yes}) or not (\emph{no}).
  \item The fact oriented checks whether the author expresses 
        objective facts (\emph{yes}) or not (\emph{no}).
}
}
\examples{
\dontrun{
#library(openxlsx)
token <- "Request your API Key: https://developer.symanto.net/"

# All actions:
#user_actions <- c("sentiments", "emotions", "action_seeking", "fact_oriented", 
#                  "self_revealing", "personality", "information_seeking")  
#excel_dest_l1 <- label_text(excel_orig, user_actions, "UTF-8", token, TRUE)
#write.xlsx(excel_dest_l1, "excel_dest_l1.xlsx")

# A sample of actions:
user_actions <- c("sentiments", "fact_oriented", "personality")  
excel_dest_l2 <- label_text(excel_orig, user_actions, "UTF-8", token, TRUE)
#write.xlsx(excel_dest_l2, "excel_dest_l2.xlsx")
}

}
\seealso{
\code{\link{get_emotion}}, \code{\link{get_sentiment}},
\code{\link{get_bua_action}}, \code{\link{get_bua_fact}},
\code{\link{get_bua_self}}, \code{\link{get_bua_personality}},
\code{\link{get_bua_info}}
}
\author{
Symanto
}
