\name{excel_orig}
\alias{excel_orig}
\docType{data}
\title{
Texts to label
}
\description{
This is an illustrative data frame with 50 posts. The first column is the text 
identifier. The second column contains the texts. The third column contains the 
language in which each text was written. 47 texts are written in English and 3 
in German.
}
\usage{excel_orig}
\format{Data frame with 50 rows and 3 columns.}
\source{\url{https://www.symanto.com/}}
\keyword{datasets}
